add some missing documentation and change pos coordinates to save as Ints

+++ src/main/kotlin/us/spaceclouds42/builders/data/spec/Pos.kt
@@ -7,9 +7,23 @@ import net.minecraft.util.Identifier
  * position of zone corners
  */
 data class Pos(
+    /**
+     * Dimension identifier
+     */
     val world: Identifier,
 
-    val x: Double,
-    val y: Double,
-    val z: Double,
+    /**
+     * x coordinate
+     */
+    val x: Int,
+
+    /**
+     * y coordinate
+     */
+    val y: Int,
+
+    /**
+     * z coordinate
+     */
+    val z: Int,
 )
